+++ crates/orchestrator/src/tests/jobs/mod.rs
@@ -18,7 +18,9 @@ use crate::jobs::types::{ExternalId, JobItem, JobStatus, JobType, JobVerificatio
 use crate::jobs::{
     create_job, handle_job_failure, increment_key_in_metadata, process_job, verify_job, Job, JobError, MockJob,
 };
-use crate::queue::job_queue::{JOB_PROCESSING_QUEUE, JOB_VERIFICATION_QUEUE};
+use crate::queue::job_queue::{
+    QueueNameForJobType, DATA_SUBMISSION_JOB_PROCESSING_QUEUE, DATA_SUBMISSION_JOB_VERIFICATION_QUEUE,
+};
 use crate::tests::common::MessagePayloadType;
 use crate::tests::config::{ConfigType, TestConfigBuilder};
 
@@ -76,7 +78,7 @@ async fn create_job_job_does_not_exists_in_db_works() {
 
     // Queue checks.
     let consumed_messages =
-        services.config.queue().consume_message_from_queue(JOB_PROCESSING_QUEUE.to_string()).await.unwrap();
+        services.config.queue().consume_message_from_queue(job_item.job_type.process_queue_name()).await.unwrap();
     let consumed_message_payload: MessagePayloadType = consumed_messages.payload_serde_json().unwrap().unwrap();
     assert_eq!(consumed_message_payload.id, job_item.id);
 }
@@ -94,7 +96,7 @@ async fn create_job_job_exists_in_db_works() {
         .await;
 
     let database_client = services.config.database();
-    database_client.create_job(job_item).await.unwrap();
+    database_client.create_job(job_item.clone()).await.unwrap();
 
     assert!(
         create_job(JobType::ProofCreation, "0".to_string(), HashMap::new(), services.config.clone()).await.is_err()
@@ -105,7 +107,7 @@ async fn create_job_job_exists_in_db_works() {
 
     // Queue checks.
     let consumed_messages =
-        services.config.queue().consume_message_from_queue(JOB_PROCESSING_QUEUE.to_string()).await.unwrap_err();
+        services.config.queue().consume_message_from_queue(job_item.job_type.process_queue_name()).await.unwrap_err();
     assert_matches!(consumed_messages, QueueError::NoData);
 }
 
@@ -125,16 +127,16 @@ async fn create_job_job_handler_is_not_implemented_panics() {
     let ctx = mock_factory::get_job_handler_context();
     ctx.expect().times(1).returning(|_| panic!("Job type not implemented yet."));
 
-    assert!(
-        create_job(JobType::ProofCreation, "0".to_string(), HashMap::new(), services.config.clone()).await.is_err()
-    );
+    let job_type = JobType::ProofCreation;
+
+    assert!(create_job(job_type.clone(), "0".to_string(), HashMap::new(), services.config.clone()).await.is_err());
 
     // Waiting for 5 secs for message to be passed into the queue
     sleep(Duration::from_secs(5)).await;
 
     // Queue checks.
     let consumed_messages =
-        services.config.queue().consume_message_from_queue(JOB_PROCESSING_QUEUE.to_string()).await.unwrap_err();
+        services.config.queue().consume_message_from_queue(job_type.process_queue_name()).await.unwrap_err();
     assert_matches!(consumed_messages, QueueError::NoData);
 }
 
@@ -184,7 +186,7 @@ async fn process_job_with_job_exists_in_db_and_valid_job_processing_status_works
 
     // Queue checks
     let consumed_messages =
-        services.config.queue().consume_message_from_queue(JOB_VERIFICATION_QUEUE.to_string()).await.unwrap();
+        services.config.queue().consume_message_from_queue(job_type.verify_queue_name()).await.unwrap();
     let consumed_message_payload: MessagePayloadType = consumed_messages.payload_serde_json().unwrap().unwrap();
     assert_eq!(consumed_message_payload.id, job_item.id);
 }
@@ -219,7 +221,7 @@ async fn process_job_with_job_exists_in_db_with_invalid_job_processing_status_er
 
     // Queue checks.
     let consumed_messages =
-        services.config.queue().consume_message_from_queue(JOB_VERIFICATION_QUEUE.to_string()).await.unwrap_err();
+        services.config.queue().consume_message_from_queue(job_item.job_type.verify_queue_name()).await.unwrap_err();
     assert_matches!(consumed_messages, QueueError::NoData);
 }
 
@@ -245,7 +247,7 @@ async fn process_job_job_does_not_exists_in_db_works() {
 
     // Queue checks.
     let consumed_messages =
-        services.config.queue().consume_message_from_queue(JOB_VERIFICATION_QUEUE.to_string()).await.unwrap_err();
+        services.config.queue().consume_message_from_queue(job_item.job_type.verify_queue_name()).await.unwrap_err();
     assert_matches!(consumed_messages, QueueError::NoData);
 }
 
@@ -380,11 +382,19 @@ async fn verify_job_with_verified_status_works() {
     sleep(Duration::from_secs(5)).await;
 
     // Queue checks.
-    let consumed_messages_verification_queue =
-        services.config.queue().consume_message_from_queue(JOB_VERIFICATION_QUEUE.to_string()).await.unwrap_err();
+    let consumed_messages_verification_queue = services
+        .config
+        .queue()
+        .consume_message_from_queue(DATA_SUBMISSION_JOB_VERIFICATION_QUEUE.to_string())
+        .await
+        .unwrap_err();
     assert_matches!(consumed_messages_verification_queue, QueueError::NoData);
-    let consumed_messages_processing_queue =
-        services.config.queue().consume_message_from_queue(JOB_PROCESSING_QUEUE.to_string()).await.unwrap_err();
+    let consumed_messages_processing_queue = services
+        .config
+        .queue()
+        .consume_message_from_queue(DATA_SUBMISSION_JOB_PROCESSING_QUEUE.to_string())
+        .await
+        .unwrap_err();
     assert_matches!(consumed_messages_processing_queue, QueueError::NoData);
 }
 
@@ -426,8 +436,12 @@ async fn verify_job_with_rejected_status_adds_to_queue_works() {
     sleep(Duration::from_secs(5)).await;
 
     // Queue checks.
-    let consumed_messages =
-        services.config.queue().consume_message_from_queue(JOB_PROCESSING_QUEUE.to_string()).await.unwrap();
+    let consumed_messages = services
+        .config
+        .queue()
+        .consume_message_from_queue(DATA_SUBMISSION_JOB_PROCESSING_QUEUE.to_string())
+        .await
+        .unwrap();
     let consumed_message_payload: MessagePayloadType = consumed_messages.payload_serde_json().unwrap().unwrap();
     assert_eq!(consumed_message_payload.id, job_item.id);
 }
@@ -478,7 +492,7 @@ async fn verify_job_with_rejected_status_works() {
 
     // Queue checks.
     let consumed_messages_processing_queue =
-        services.config.queue().consume_message_from_queue(JOB_PROCESSING_QUEUE.to_string()).await.unwrap_err();
+        services.config.queue().consume_message_from_queue(job_item.job_type.process_queue_name()).await.unwrap_err();
     assert_matches!(consumed_messages_processing_queue, QueueError::NoData);
 }
 
@@ -524,7 +538,7 @@ async fn verify_job_with_pending_status_adds_to_queue_works() {
 
     // Queue checks
     let consumed_messages =
-        services.config.queue().consume_message_from_queue(JOB_VERIFICATION_QUEUE.to_string()).await.unwrap();
+        services.config.queue().consume_message_from_queue(job_item.job_type.verify_queue_name()).await.unwrap();
     let consumed_message_payload: MessagePayloadType = consumed_messages.payload_serde_json().unwrap().unwrap();
     assert_eq!(consumed_message_payload.id, job_item.id);
 }
@@ -576,7 +590,7 @@ async fn verify_job_with_pending_status_works() {
 
     // Queue checks.
     let consumed_messages_verification_queue =
-        services.config.queue().consume_message_from_queue(JOB_VERIFICATION_QUEUE.to_string()).await.unwrap_err();
+        services.config.queue().consume_message_from_queue(job_item.job_type.verify_queue_name()).await.unwrap_err();
     assert_matches!(consumed_messages_verification_queue, QueueError::NoData);
 }