+++ crates/orchestrator/src/tests/workers/proving/mod.rs
@@ -15,6 +15,7 @@ use crate::database::MockDatabase;
 use crate::jobs::job_handler_factory::mock_factory;
 use crate::jobs::types::{JobItem, JobStatus, JobType};
 use crate::jobs::{Job, MockJob};
+use crate::queue::job_queue::PROVING_JOB_PROCESSING_QUEUE;
 use crate::queue::MockQueueProvider;
 use crate::tests::config::TestConfigBuilder;
 use crate::tests::workers::utils::{db_checks_proving_worker, get_job_by_mock_id_vector};
@@ -31,8 +32,6 @@ async fn test_proving_worker(#[case] incomplete_runs: bool) -> Result<(), Box<dy
     let mut queue = MockQueueProvider::new();
     let prover_client = MockProverClient::new();
     let settlement_client = MockSettlementClient::new();
-
-    const JOB_PROCESSING_QUEUE: &str = "madara_orchestrator_job_processing_queue";
 
     // Mocking Prover Client
 
@@ -63,7 +62,7 @@ async fn test_proving_worker(#[case] incomplete_runs: bool) -> Result<(), Box<dy
             .expect_send_message_to_queue()
             .times(4)
             .returning(|_, _, _| Ok(()))
-            .withf(|queue, _payload, _delay| queue == JOB_PROCESSING_QUEUE);
+            .withf(|queue, _payload, _delay| queue == PROVING_JOB_PROCESSING_QUEUE);
     } else {
         for i in 1..5 + 1 {
             db_checks_proving_worker(i, &mut db, &mut job_handler);
@@ -80,7 +79,7 @@ async fn test_proving_worker(#[case] incomplete_runs: bool) -> Result<(), Box<dy
             .expect_send_message_to_queue()
             .times(5)
             .returning(|_, _, _| Ok(()))
-            .withf(|queue, _payload, _delay| queue == JOB_PROCESSING_QUEUE);
+            .withf(|queue, _payload, _delay| queue == PROVING_JOB_PROCESSING_QUEUE);
     }
 
     let provider = JsonRpcClient::new(HttpTransport::new(