+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Added
 
+- Add multiple queues for processing and verification based on job type
 - added logs
 - added MongoDB migrations using nodejs
 - added dockerfile